requirements.txt
beautifulsoup4==4.5.1
chardet==2.3.0
click==6.6
cssselect==1.0.0
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
lxml==3.7.1
MarkupSafe==0.23
readability-lxml==0.6.2
requests==2.12.4
Werkzeug==0.11.13
wheel==0.24.0
